修正儲值卡payment 欄位型態錯誤

+++ app/Models/PrepaidCardRecord.php
@@ -38,8 +38,8 @@ class PrepaidCardRecord extends Model
         'member_id' => 'string',
         'branch_id' => 'string',
         'type' => 'string',
-        'topup' => 'integer',
-        'payment' => 'integer',
+        'topup' => 'double',
+        'payment' => 'double',
         'void_id' => 'string',
         'remark' => 'string'
     ];
